feat: add routines to update total and circulating supply
Check if supply or circulating columns are empty for any one token in existing tokens stored in the database.
Add total and circulating if missing.
Check if the network returns a different value for any total or circulating compared to those stored in the database.
Update total and circulating if different.

fix - Remove the commented-out scheduler runner from the previous release.